feat(p11): durable job queue + UI wiring + polling hardening
Client-side SQLite job queue (src-tauri/src/jobs/):
- JobQueue with two-table schema: ingest_jobs + file_queue (WAL mode,
  bundled rusqlite). Per-file status pending|in_progress|done|error|skipped,
  INSERT OR IGNORE dedup on (job_id, path), claim_batch with BEGIN IMMEDIATE,
  reclaim_in_progress for crash recovery.
- 16 Tauri commands: jobs_create (optional job_type), jobs_list, jobs_get,
  jobs_set_status, jobs_delete, jobs_add_files, jobs_claim_batch,
  jobs_mark_done, jobs_mark_error, jobs_mark_skipped, jobs_set_doc_id,
  jobs_reclaim, jobs_pending_count, jobs_list_files, jobs_remove_file,
  jobs_remove_files_by_status. All async via spawn_blocking.

IndexIngest.svelte wired to durable queue:
- On mount: jobs_list → find active hinzufuegen job → jobs_reclaim →
  jobs_list_files to restore display (replaces tauri-plugin-store entries).
- addPaths → jobs_create (job_type='hinzufuegen') + jobs_add_files.
- L1/L2/L3 ingest loops use jobs_claim_batch → process → jobs_mark_done/error/skipped.
- stopIngest abort: finally block calls jobs_reclaim + reload.
- clearAll → jobs_delete; clearDone → jobs_remove_files_by_status.
- persistEntries() removed; updateEntry no longer fires store writes.

mtime_unix Option<u32> → Option<i64> (Y2038) across ingest.rs,
bg_ingest/mod.rs, tauri_commands.rs, local_index.rs.

Remote batch ingest hardening:
- Adaptive task-status polling: 500ms×4 → 2s×4 → 5s (instead of
  hardcoded 500ms), saving ~10× IPC calls on long server batches.
- Server queue depth chip: $effect gates setInterval on
  batchManager.isProcessing instead of polling 24/7.
- crisp-index-server ingest_batch + task_status handlers now wrap
  synchronous SQLite calls in tokio::task::spawn_blocking (issue D).

Server-side (crisp-index-server/src/queue.rs + embedder.rs):
- TaskQueue with lease/heartbeat/retry, reclaim_expired_tasks,
  legacy JSON migration. enqueue_batch + get_task_status moved off
  the Tokio thread pool via spawn_blocking in their Axum handlers.
- ServerEmbedder (fastembed BGE-M3) hydrates missing embeddings.

+++ src-tauri/src/index/local_index.rs
@@ -602,7 +602,7 @@ impl LocalIndex {
     /// Background ingest uses this to mtime-skip files that haven't
     /// changed since last index — saves the read + extract + embed
     /// cost on the common "no new content" case.
-    pub async fn indexed_mtime_for_uri(&self, location_uri: &str) -> Result<Option<u32>> {
+    pub async fn indexed_mtime_for_uri(&self, location_uri: &str) -> Result<Option<i64>> {
         let pred = format!(
             "location_uri = '{}' AND chunk_index = 0",
             location_uri.replace('\'', "''")
@@ -640,7 +640,7 @@ impl LocalIndex {
                         // Read the integer up to the next non-digit.
                         let end = rest.find(|c: char| !c.is_ascii_digit()).unwrap_or(rest.len());
                         if end > 0 {
-                            if let Ok(v) = rest[..end].parse::<u32>() {
+                            if let Ok(v) = rest[..end].parse::<i64>() {
                                 return Ok(Some(v));
                             }
                         }
